+++ packages/core/src/errors/RequestError/index.ts
@@ -1,9 +1,7 @@
 import { LogtoErrorCode, LogtoErrorI18nKey } from '@logto/phrases';
-import { RequestErrorBody, RequestErrorMetadata, UserLogDBEntry } from '@logto/schemas';
+import { RequestErrorBody, RequestErrorMetadata } from '@logto/schemas';
 import i18next from 'i18next';
 import pick from 'lodash.pick';
-
-import { OmitAutoSetFields } from '@/database/utils';
 
 export default class RequestError extends Error {
   code: LogtoErrorCode;
@@ -29,18 +27,5 @@ export default class RequestError extends Error {
 
   get body(): RequestErrorBody {
     return pick(this, 'message', 'code', 'data');
-  }
-}
-
-export class RequestErrorWithUserLog extends RequestError {
-  payload: OmitAutoSetFields<UserLogDBEntry>;
-
-  constructor(
-    input: RequestErrorMetadata | LogtoErrorCode,
-    payload: OmitAutoSetFields<UserLogDBEntry>,
-    data?: unknown
-  ) {
-    super(input, data);
-    this.payload = payload;
   }
 }
